+++ shuup/xtheme/layout/_base.py
@@ -10,6 +10,7 @@ from __future__ import unicode_literals
 from django.utils.safestring import mark_safe
 from django.utils.translation import ugettext_lazy as _
 
+from shuup.core import cache
 from shuup.utils.django_compat import force_text
 from shuup.xtheme.plugins._base import Plugin
 
@@ -72,7 +73,7 @@ class LayoutCell(object):
             return plugin_class(config=self.config)
         return None
 
-    def render(self, context):
+    def render(self, context, cache_key_prefix=None):
         """
         Return the plugin's rendered content.
 
@@ -86,8 +87,20 @@ class LayoutCell(object):
         plugin_inst = self.instantiate_plugin()
         if plugin_inst is None:
             return mark_safe("<!-- %s? -->" % self.plugin_identifier)
+
         if plugin_inst.is_context_valid(context=context):
-            return plugin_inst.render(context=context)
+            # check whether the plugin can be cached
+            cacheabled = getattr(plugin_inst, "cacheable", False)
+            cache_key = plugin_inst.get_cache_key() if hasattr(plugin_inst, "get_cache_key") else plugin_inst.identifier
+            full_cache_key = f"{cache_key_prefix}|{cache_key}"
+            cached_content = cache.get(full_cache_key)
+            if cached_content is not None:
+                return cached_content
+
+            content = plugin_inst.render(context=context)
+            if cacheabled:
+                cache.set(full_cache_key, content)
+            return content
         else:
             return ""